+++ Changelog.md
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+### Version 3.1.0 (2021-09-17)
+* Moving all the Entities interfaces in bundle entity-bundle
+* Adding manage Library with multi domain
+* Removed entity-seo-bundle
+* Adding new bundle seo-bundle
+* Added ModuleListener for filter admin list
+
 ### Version 3.0.0 (2021-02-10)
 * Created new repository and new branch in to Github
 * Changed the project License to the MIT